moved rolling wrapper function out of stats into utils
